# frontend/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "carHelper server",
		"description": "\nA tool to help you manage your car's service history, expenses, mileage, and more.\n",
		"contact": { "name": "carHelper", "url": "https://rishimo.dev/", "email": "hello@rishimo.dev" },
		"version": "0.1.0"
	},
	"paths": {
...
		"/user/vehicles": {
			"get": {
				"tags": ["user"],
				"summary": "Get Vehicles",
				"description": "Return the user's vehicles.",
				"operationId": "get_vehicles_user_vehicles_get",
				"security": [{ "JwtAccessBearer": [] }],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": { "application/json": { "schema": {} } }
					}
				}
			},
			"post": {
				"tags": ["user"],
				"summary": "Add Vehicle",
				"description": "Add a vehicle to the user's vehicles.",
				"operationId": "add_vehicle_user_vehicles_post",
				"security": [{ "JwtAccessBearer": [] }],
				"parameters": [
					{
						"name": "vehicle_id",
						"in": "query",
						"required": true,
						"schema": { "type": "string", "title": "Vehicle Id" }
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": { "application/json": { "schema": {} } }
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["user"],
				"summary": "Remove Vehicle",
				"description": "Remove a vehicle from the user's vehicles.",
				"operationId": "remove_vehicle_user_vehicles_delete",
				"security": [{ "JwtAccessBearer": [] }],
				"parameters": [
					{
						"name": "vehicle_id",
						"in": "query",
						"required": true,
						"schema": { "type": "string", "title": "Vehicle Id" }
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": { "application/json": { "schema": {} } }
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
...
		"/vehicle": {
			"post": {
				"tags": ["vehicle"],
				"summary": "Create Vehicle",
				"description": "Create a new vehicle and associate it with the user.",
				"operationId": "create_vehicle_vehicle_post",
				"requestBody": {
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/Vehicle-Input" } }
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/Vehicle-Output" } }
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				},
				"security": [{ "JwtAccessBearer": [] }]
			}
		},
		"/vehicle/my": {
			"get": {
				"tags": ["vehicle"],
				"summary": "Get My Vehicles",
				"description": "Get all vehicles belonging to the current user.",
				"operationId": "get_my_vehicles_vehicle_my_get",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"items": { "$ref": "#/components/schemas/Vehicle-Output" },
									"type": "array",
									"title": "Response Get My Vehicles Vehicle My Get"
								}
							}
						}
					}
				},
				"security": [{ "JwtAccessBearer": [] }]
			}
		},
		"/vehicle/{vehicle_id}": {
			"get": {
				"tags": ["vehicle"],
				"summary": "Get Vehicle",
				"description": "Get a specific vehicle by ID.",
				"operationId": "get_vehicle_vehicle__vehicle_id__get",
				"security": [{ "JwtAccessBearer": [] }],
				"parameters": [
					{
						"name": "vehicle_id",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "title": "Vehicle Id" }
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/Vehicle-Output" } }
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"patch": {
				"tags": ["vehicle"],
				"summary": "Update Vehicle",
				"description": "Update a vehicle's details.",
				"operationId": "update_vehicle_vehicle__vehicle_id__patch",
				"security": [{ "JwtAccessBearer": [] }],
				"parameters": [
					{
						"name": "vehicle_id",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "title": "Vehicle Id" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": { "schema": { "type": "object", "title": "Update Data" } }
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/Vehicle-Output" } }
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["vehicle"],
				"summary": "Delete Vehicle",
				"description": "Delete a vehicle and remove it from user's vehicles.",
				"operationId": "delete_vehicle_vehicle__vehicle_id__delete",
				"security": [{ "JwtAccessBearer": [] }],
				"parameters": [
					{
						"name": "vehicle_id",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "title": "Vehicle Id" }
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": { "application/json": { "schema": {} } }
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/vehicle/{vehicle_id}/stats": {
			"get": {
				"tags": ["vehicle"],
				"summary": "Get Vehicle Stats",
				"description": "Get comprehensive stats for a vehicle.",
				"operationId": "get_vehicle_stats_vehicle__vehicle_id__stats_get",
				"security": [{ "JwtAccessBearer": [] }],
				"parameters": [
					{
						"name": "vehicle_id",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "title": "Vehicle Id" }
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": { "application/json": { "schema": {} } }
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/vehicle/{vehicle_id}/odometer": {
			"post": {
				"tags": ["vehicle"],
				"summary": "Add Odometer Reading",
				"description": "Add a new odometer reading for a vehicle.",
				"operationId": "add_odometer_reading_vehicle__vehicle_id__odometer_post",
				"security": [{ "JwtAccessBearer": [] }],
				"parameters": [
					{
						"name": "vehicle_id",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "title": "Vehicle Id" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/OdometerReading" } }
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": { "application/json": { "schema": {} } }
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/vehicle/{vehicle_id}/service": {
			"post": {
				"tags": ["vehicle"],
				"summary": "Add Service Record",
				"description": "Add a service record to a vehicle.",
				"operationId": "add_service_record_vehicle__vehicle_id__service_post",
				"security": [{ "JwtAccessBearer": [] }],
				"parameters": [
					{
						"name": "vehicle_id",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "title": "Vehicle Id" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/Service-Input" } }
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/Service-Output" } }
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/vehicle/{vehicle_id}/fuel": {
			"post": {
				"tags": ["vehicle"],
				"summary": "Add Fuel Record",
				"description": "Add a fuel record to a vehicle.",
				"operationId": "add_fuel_record_vehicle__vehicle_id__fuel_post",
				"security": [{ "JwtAccessBearer": [] }],
				"parameters": [
					{
						"name": "vehicle_id",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "title": "Vehicle Id" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/Fuel-Input" } }
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/Fuel-Output" } }
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/vehicle/{vehicle_id}/files": {
			"post": {
				"tags": ["vehicle"],
				"summary": "Upload File",
				"description": "Upload a file associated with the vehicle.",
				"operationId": "upload_file_vehicle__vehicle_id__files_post",
				"security": [{ "JwtAccessBearer": [] }],
				"parameters": [
					{
						"name": "vehicle_id",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "title": "Vehicle Id" }
					},
					{
						"name": "description",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [{ "type": "string" }, { "type": "null" }],
							"title": "Description"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"$ref": "#/components/schemas/Body_upload_file_vehicle__vehicle_id__files_post"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": { "application/json": { "schema": {} } }
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
...
			"Body_upload_file_vehicle__vehicle_id__files_post": {
				"properties": { "file": { "type": "string", "format": "binary", "title": "File" } },
				"type": "object",
				"required": ["file"],
				"title": "Body_upload_file_vehicle__vehicle_id__files_post"
			},
			"Expense-Input": {
				"properties": {
					"_id": { "type": "string", "title": "Id", "description": "Document ID" },
					"VIN": {
						"anyOf": [{ "type": "string" }, { "type": "null" }],
						"title": "Vin",
						"description": "Vehicle VIN"
					},
					"date": { "type": "string", "format": "date-time", "title": "Date" },
					"vendor": { "$ref": "#/components/schemas/Vendor", "description": "Expense vendor" },
					"description": {
						"type": "string",
						"title": "Description",
						"description": "Expense description"
					},
					"cost": { "type": "number", "title": "Cost", "description": "Expense cost" },
					"file_ids": {
						"items": { "type": "string" },
						"type": "array",
						"title": "File Ids",
						"description": "File IDs for the expense"
					},
					"items": {
						"items": { "$ref": "#/components/schemas/ExpenseItem" },
						"type": "array",
						"title": "Items",
						"description": "Expense items"
					}
				},
				"type": "object",
				"required": ["_id", "vendor", "description", "cost"],
				"title": "Expense",
				"description": "Expense record, for any other expenses like parts, mods, tools, etc."
			},
			"Expense-Output": {
				"properties": {
					"_id": { "type": "string", "title": "Id", "description": "Document ID" },
					"VIN": {
						"anyOf": [{ "type": "string" }, { "type": "null" }],
						"title": "Vin",
						"description": "Vehicle VIN"
					},
					"date": { "type": "string", "format": "date-time", "title": "Date" },
					"vendor": { "$ref": "#/components/schemas/Vendor", "description": "Expense vendor" },
					"description": {
						"type": "string",
						"title": "Description",
						"description": "Expense description"
					},
					"cost": { "type": "number", "title": "Cost", "description": "Expense cost" },
					"file_ids": {
						"items": { "type": "string" },
						"type": "array",
						"title": "File Ids",
						"description": "File IDs for the expense"
					},
					"items": {
						"items": { "$ref": "#/components/schemas/ExpenseItem" },
						"type": "array",
						"title": "Items",
						"description": "Expense items"
					}
				},
				"type": "object",
				"required": ["_id", "vendor", "description", "cost"],
				"title": "Expense",
				"description": "Expense record, for any other expenses like parts, mods, tools, etc."
			},
...
			"Fuel-Input": {
				"properties": {
					"_id": { "type": "string", "title": "Id", "description": "Document ID" },
					"VIN": { "type": "string", "title": "Vin", "description": "Vehicle VIN" },
					"date": { "type": "string", "format": "date-time", "title": "Date" },
					"vendor_id": { "type": "string", "title": "Vendor Id", "description": "Fuel vendor ID" },
					"fuel_type": {
						"$ref": "#/components/schemas/FuelType",
						"description": "Type of fuel purchased"
					},
					"units": { "type": "number", "title": "Units", "description": "Units of fuel purchased" },
					"price_per_unit": {
						"type": "number",
						"title": "Price Per Unit",
						"description": "Price per unit of fuel"
					},
					"cost": {
						"anyOf": [{ "type": "number" }, { "type": "null" }],
						"title": "Cost",
						"description": "Total cost of the fuel purchase"
					},
					"attachments": {
						"items": { "$ref": "#/components/schemas/File" },
						"type": "array",
						"title": "Attachments",
						"description": "Documentation for the fuel purchase"
					}
				},
				"type": "object",
				"required": ["_id", "VIN", "vendor_id", "fuel_type", "units", "price_per_unit"],
				"title": "Fuel",
				"description": "Fuel purchase record\n\nAttributes:\n    VIN (str): Vehicle Identification Number\n    date (DateTime): Date and time of fuel purchase\n    vendor_id (str): Fuel vendor ID\n    fuel_type (FuelType): Type of fuel purchased\n    units (float): Amount of fuel purchased\n    price_per_unit (float): Cost per unit of fuel\n    total_cost (float): Total cost of the fuel purchase"
			},
			"Fuel-Output": {
				"properties": {
					"_id": { "type": "string", "title": "Id", "description": "Document ID" },
					"VIN": { "type": "string", "title": "Vin", "description": "Vehicle VIN" },
					"date": { "type": "string", "format": "date-time", "title": "Date" },
					"vendor_id": { "type": "string", "title": "Vendor Id", "description": "Fuel vendor ID" },
					"fuel_type": {
						"$ref": "#/components/schemas/FuelType",
						"description": "Type of fuel purchased"
					},
					"units": { "type": "number", "title": "Units", "description": "Units of fuel purchased" },
					"price_per_unit": {
						"type": "number",
						"title": "Price Per Unit",
						"description": "Price per unit of fuel"
					},
					"cost": {
						"anyOf": [{ "type": "number" }, { "type": "null" }],
						"title": "Cost",
						"description": "Total cost of the fuel purchase"
					},
					"attachments": {
						"items": { "$ref": "#/components/schemas/File" },
						"type": "array",
						"title": "Attachments",
						"description": "Documentation for the fuel purchase"
					}
				},
				"type": "object",
				"required": ["_id", "VIN", "vendor_id", "fuel_type", "units", "price_per_unit"],
				"title": "Fuel",
				"description": "Fuel purchase record\n\nAttributes:\n    VIN (str): Vehicle Identification Number\n    date (DateTime): Date and time of fuel purchase\n    vendor_id (str): Fuel vendor ID\n    fuel_type (FuelType): Type of fuel purchased\n    units (float): Amount of fuel purchased\n    price_per_unit (float): Cost per unit of fuel\n    total_cost (float): Total cost of the fuel purchase"
			},
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": { "$ref": "#/components/schemas/ValidationError" },
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
			"Location": {
				"properties": {
					"address": { "type": "string", "title": "Address" },
					"city": { "type": "string", "title": "City" },
					"state": { "type": "string", "title": "State" },
					"zip_code": { "type": "string", "title": "Zip Code" },
					"latitude": {
						"anyOf": [{ "type": "number" }, { "type": "null" }],
						"title": "Latitude",
						"description": "Latitude of the location"
					},
					"longitude": {
						"anyOf": [{ "type": "number" }, { "type": "null" }],
						"title": "Longitude",
						"description": "Longitude of the location"
					}
				},
				"type": "object",
				"required": ["address", "city", "state", "zip_code"],
				"title": "Location",
				"description": "Location model"
			},
			"OdometerReading": {
				"properties": {
					"_id": { "type": "string", "title": "Id", "description": "Document ID" },
					"VIN": { "type": "string", "title": "Vin", "description": "Vehicle VIN" },
					"date": { "type": "string", "format": "date-time", "title": "Date" },
					"reading": { "type": "number", "title": "Reading" },
					"location": {
						"anyOf": [{ "$ref": "#/components/schemas/Location" }, { "type": "null" }],
						"description": "Odometer reading location"
					}
				},
				"type": "object",
				"required": ["_id", "VIN", "reading"],
				"title": "OdometerReading",
				"description": "Vehicle odometer reading record\n\nAttributes:\n    VIN (str): Vehicle Identification Number\n    date (DateTime): Date and time of the reading\n    reading (float): Odometer value\n    location (Optional[Location]): Location where reading was taken"
			},
			"Service-Input": {
				"properties": {
					"_id": { "type": "string", "title": "Id", "description": "Document ID" },
					"VIN": { "type": "string", "title": "Vin", "description": "Vehicle VIN" },
					"date": { "type": "string", "format": "date-time", "title": "Date" },
					"vendor_id": {
						"type": "string",
						"title": "Vendor Id",
						"description": "Service vendor ID"
					},
					"cost": {
						"anyOf": [{ "type": "number" }, { "type": "null" }],
						"title": "Cost",
						"description": "Total cost of all service items"
					},
					"items": {
						"items": { "$ref": "#/components/schemas/ServiceItem-Input" },
						"type": "array",
						"title": "Items"
					}
				},
				"type": "object",
				"required": ["_id", "VIN", "vendor_id", "items"],
				"title": "Service",
				"description": "Complete service record for a vehicle service visit\n\nAttributes:\n    VIN (str): Vehicle Identification Number\n    date (DateTime): Date and time of the service\n    vendor_id (str): Service provider ID\n    total_cost (float): Total cost of all service items\n    items (List[ServiceItem]): List of individual service items performed"
			},
			"Service-Output": {
				"properties": {
					"_id": { "type": "string", "title": "Id", "description": "Document ID" },
					"VIN": { "type": "string", "title": "Vin", "description": "Vehicle VIN" },
					"date": { "type": "string", "format": "date-time", "title": "Date" },
					"vendor_id": {
						"type": "string",
						"title": "Vendor Id",
						"description": "Service vendor ID"
					},
					"cost": {
						"anyOf": [{ "type": "number" }, { "type": "null" }],
						"title": "Cost",
						"description": "Total cost of all service items"
					},
					"items": {
						"items": { "$ref": "#/components/schemas/ServiceItem-Output" },
						"type": "array",
						"title": "Items"
					}
				},
				"type": "object",
				"required": ["_id", "VIN", "vendor_id", "items"],
				"title": "Service",
				"description": "Complete service record for a vehicle service visit\n\nAttributes:\n    VIN (str): Vehicle Identification Number\n    date (DateTime): Date and time of the service\n    vendor_id (str): Service provider ID\n    total_cost (float): Total cost of all service items\n    items (List[ServiceItem]): List of individual service items performed"
			},
			"ServiceItem-Input": {
				"properties": {
					"_id": { "type": "string", "title": "Id", "description": "Document ID" },
					"service_type": {
						"$ref": "#/components/schemas/ServiceType",
						"description": "Service type",
						"default": "MAINTENANCE"
					},
					"description": {
						"type": "string",
						"title": "Description",
						"description": "Service description"
					},
					"cost": { "type": "number", "title": "Cost", "description": "Service cost" },
					"attachments": {
						"items": { "$ref": "#/components/schemas/File" },
						"type": "array",
						"title": "Attachments",
						"description": "Documentation for the service"
					}
				},
				"type": "object",
				"required": ["_id", "description", "cost"],
				"title": "ServiceItem",
				"description": "Individual service item or task performed during a service visit\n\nAttributes:\n    service_type (ServiceType): Type of service performed (e.g., MAINTENANCE)\n    description (str): Detailed description of the service performed\n    cost (float): Cost of the individual service item\n    attachments (List[File]): Related documentation for the service item"
			},
...
			"ValidationError": {
				"properties": {
					"loc": {
						"items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
						"type": "array",
						"title": "Location"
					},
					"msg": { "type": "string", "title": "Message" },
					"type": { "type": "string", "title": "Error Type" }
				},
				"type": "object",
				"required": ["loc", "msg", "type"],
				"title": "ValidationError"
			},
			"Vehicle-Input": {
				"properties": {
					"_id": { "type": "string", "title": "Id", "description": "Document ID" },
					"VIN": { "type": "string", "title": "Vin", "description": "Vehicle VIN" },
					"make": { "type": "string", "title": "Make", "description": "Vehicle make" },
					"model": { "type": "string", "title": "Model", "description": "Vehicle model" },
					"year": { "type": "integer", "title": "Year", "description": "Vehicle year" },
					"color": {
						"anyOf": [{ "type": "string" }, { "type": "null" }],
						"title": "Color",
						"description": "Vehicle color",
						"default": "NOT_SET"
					},
					"owner_id": {
						"anyOf": [{ "type": "string" }, { "type": "null" }],
						"title": "Owner Id",
						"description": "Vehicle owner ID"
					},
					"purchase_date": {
						"anyOf": [{ "type": "string", "format": "date-time" }, { "type": "null" }],
						"title": "Purchase Date",
						"description": "Vehicle purchase date"
					},
					"purchase_price": {
						"anyOf": [{ "type": "number" }, { "type": "null" }],
						"title": "Purchase Price",
						"description": "Vehicle purchase price"
					},
					"odometer_records": {
						"anyOf": [
							{ "items": { "$ref": "#/components/schemas/OdometerReading" }, "type": "array" },
							{ "type": "null" }
						],
						"title": "Odometer Records",
						"description": "Odometer records"
					},
					"service_records": {
						"anyOf": [
							{ "items": { "$ref": "#/components/schemas/Service-Input" }, "type": "array" },
							{ "type": "null" }
						],
						"title": "Service Records",
						"description": "Service records"
					},
					"fuel_records": {
						"anyOf": [
							{ "items": { "$ref": "#/components/schemas/Fuel-Input" }, "type": "array" },
							{ "type": "null" }
						],
						"title": "Fuel Records",
						"description": "Fuel records"
					},
					"expenses": {
						"anyOf": [
							{ "items": { "$ref": "#/components/schemas/Expense-Input" }, "type": "array" },
							{ "type": "null" }
						],
						"title": "Expenses",
						"description": "Expenses"
					},
					"file_ids": {
						"anyOf": [{ "items": { "type": "string" }, "type": "array" }, { "type": "null" }],
						"title": "File Ids",
						"description": "File IDs"
					}
				},
				"type": "object",
				"required": ["_id", "VIN", "make", "model", "year"],
				"title": "Vehicle",
				"description": "Primary vehicle record containing all vehicle-related information\n\nAttributes:\n    VIN (str): Vehicle Identification Number\n    make (str): Vehicle manufacturer\n    model (str): Vehicle model name\n    year (int): Vehicle manufacturing year\n    color (Optional[str]): Vehicle color, defaults to \"NOT_SET\"\n    owner_id (Optional[str]): Vehicle owner ID\n    odometer_records (Optional[List[OdometerReading]]): History of odometer readings\n    service_records (Optional[List[Service]]): History of service visits\n    fuel_records (Optional[List[Fuel]]): History of fuel purchases\n    documentation (Optional[List[Documentation]]): Vehicle-related documentation"
			},
			"Vehicle-Output": {
				"properties": {
					"_id": { "type": "string", "title": "Id", "description": "Document ID" },
					"VIN": { "type": "string", "title": "Vin", "description": "Vehicle VIN" },
					"make": { "type": "string", "title": "Make", "description": "Vehicle make" },
					"model": { "type": "string", "title": "Model", "description": "Vehicle model" },
					"year": { "type": "integer", "title": "Year", "description": "Vehicle year" },
					"color": {
						"anyOf": [{ "type": "string" }, { "type": "null" }],
						"title": "Color",
						"description": "Vehicle color",
						"default": "NOT_SET"
					},
					"owner_id": {
						"anyOf": [{ "type": "string" }, { "type": "null" }],
						"title": "Owner Id",
						"description": "Vehicle owner ID"
					},
					"purchase_date": {
						"anyOf": [{ "type": "string", "format": "date-time" }, { "type": "null" }],
						"title": "Purchase Date",
						"description": "Vehicle purchase date"
					},
					"purchase_price": {
						"anyOf": [{ "type": "number" }, { "type": "null" }],
						"title": "Purchase Price",
						"description": "Vehicle purchase price"
					},
					"odometer_records": {
						"anyOf": [
							{ "items": { "$ref": "#/components/schemas/OdometerReading" }, "type": "array" },
							{ "type": "null" }
						],
						"title": "Odometer Records",
						"description": "Odometer records"
					},
					"service_records": {
						"anyOf": [
							{ "items": { "$ref": "#/components/schemas/Service-Output" }, "type": "array" },
							{ "type": "null" }
						],
						"title": "Service Records",
						"description": "Service records"
					},
					"fuel_records": {
						"anyOf": [
							{ "items": { "$ref": "#/components/schemas/Fuel-Output" }, "type": "array" },
							{ "type": "null" }
						],
						"title": "Fuel Records",
						"description": "Fuel records"
					},
					"expenses": {
						"anyOf": [
							{ "items": { "$ref": "#/components/schemas/Expense-Output" }, "type": "array" },
							{ "type": "null" }
						],
						"title": "Expenses",
						"description": "Expenses"
					},
					"file_ids": {
						"anyOf": [{ "items": { "type": "string" }, "type": "array" }, { "type": "null" }],
						"title": "File Ids",
						"description": "File IDs"
					}
				},
				"type": "object",
				"required": ["_id", "VIN", "make", "model", "year"],
				"title": "Vehicle",
				"description": "Primary vehicle record containing all vehicle-related information\n\nAttributes:\n    VIN (str): Vehicle Identification Number\n    make (str): Vehicle manufacturer\n    model (str): Vehicle model name\n    year (int): Vehicle manufacturing year\n    color (Optional[str]): Vehicle color, defaults to \"NOT_SET\"\n    owner_id (Optional[str]): Vehicle owner ID\n    odometer_records (Optional[List[OdometerReading]]): History of odometer readings\n    service_records (Optional[List[Service]]): History of service visits\n    fuel_records (Optional[List[Fuel]]): History of fuel purchases\n    documentation (Optional[List[Documentation]]): Vehicle-related documentation"
			},
...
		},
		"securitySchemes": {
			"JwtRefreshBearer": { "type": "http", "scheme": "bearer" },
			"JwtAccessBearer": { "type": "http", "scheme": "bearer" }
		}
	}
}